Properly dropped database from a series of tests

+++ console/src/test/java/com/arcadedb/console/BaseGraphServerTest.java
@@ -308,10 +308,16 @@ public abstract class BaseGraphServerTest {
   }
 
   protected void deleteDatabaseFolders() {
-    for (int i = 0; i < getServerCount(); ++i) {
-      if (getServer(i).existsDatabase(getDatabaseName()))
-        getServer(i).getDatabase(getDatabaseName()).drop();
-    }
+    if (databases != null)
+      for (int i = 0; i < databases.length; ++i) {
+        databases[i].drop();
+      }
+
+    if (servers != null)
+      for (int i = 0; i < getServerCount(); ++i) {
+        if (getServer(i).existsDatabase(getDatabaseName()))
+          getServer(i).getDatabase(getDatabaseName()).drop();
+      }
 
     for (int i = 0; i < getServerCount(); ++i)
       FileUtils.deleteRecursively(new File(getDatabasePath(i)));
