+++ server/src/test/java/performance/BasePerformanceTest.java
@@ -175,9 +175,15 @@ public abstract class BasePerformanceTest {
   }
 
   protected void deleteDatabaseFolders() {
-    for (int i = 0; i < getServerCount(); ++i)
-      if (getServer(i).existsDatabase(getDatabaseName()))
-        getServer(i).getDatabase(getDatabaseName()).drop();
+    if (databases != null)
+      for (int i = 0; i < databases.length; ++i) {
+        databases[i].drop();
+      }
+
+    if (servers != null)
+      for (int i = 0; i < getServerCount(); ++i)
+        if (getServer(i).existsDatabase(getDatabaseName()))
+          getServer(i).getDatabase(getDatabaseName()).drop();
 
     for (int i = 0; i < getServerCount(); ++i)
       FileUtils.deleteRecursively(new File(getDatabasePath(i)));
